requirements.txt
alembic==1.15.2
asgiref==3.8.1
asyncpg==0.26.0
bayoo-docx==0.2.20
celery==5.5.1
django_celery_beat==2.8.0
fastapi-jwt==0.3.0
fastapi==0.115.12
httpx==0.28.1
passlib==1.7.4
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic[email]>=2.0.3
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
redis==5.2.1
redis[asyncio]>=4.0.0
uvicorn==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
psycopg2==2.9.10
jinja2==3.1.6
